Separate the low/high level lib stuff.
