+++ lib/core/base_wrapper.rb
