+++ lib/core/dns_wrapper.rb
