+++ lib/core/ec2_wrapper.rb
