+++ lib/core/elb_wrapper.rb
