+++ lib/util/ec2_elb_starter.rb
