+++ lib/util/elb_swapper.rb
