+++ spec/dns_wrapper_spec.rb
@@ -1,4 +1,4 @@
-require_relative '../lib/dns_wrapper'
+require_relative '../lib/core/dns_wrapper'
 
 describe DnsWrapper do
   # TODO
